+++ tomcat/src/main/java/org/apache/coyote/http11/handler/Handler.java
@@ -1,7 +1,7 @@
 package org.apache.coyote.http11.handler;
 
-import org.apache.coyote.http11.request.HttpRequest;
-import org.apache.coyote.http11.response.HttpResponse;
+import org.apache.coyote.http11.http.request.HttpRequest;
+import org.apache.coyote.http11.http.response.HttpResponse;
 
 public interface Handler {
     HttpResponse handle(final HttpRequest httpRequest);
