+++ tomcat/src/main/java/org/apache/coyote/http11/handler/HomeHandler.java
@@ -1,15 +1,15 @@
 package org.apache.coyote.http11.handler;
 
-import static org.apache.coyote.http11.HttpVersion.HTTP11;
+import static org.apache.coyote.http11.http.HttpVersion.HTTP11;
 import static org.apache.coyote.http11.header.ContentType.HTML;
 import static org.apache.coyote.http11.header.ContentType.UTF_8;
 import static org.apache.coyote.http11.header.HttpHeaderType.CONTENT_LENGTH;
 import static org.apache.coyote.http11.header.HttpHeaderType.CONTENT_TYPE;
-import static org.apache.coyote.http11.response.HttpStatus.OK;
+import static org.apache.coyote.http11.http.response.HttpStatus.OK;
 
 import org.apache.coyote.http11.header.HttpHeader;
-import org.apache.coyote.http11.request.HttpRequest;
-import org.apache.coyote.http11.response.HttpResponse;
+import org.apache.coyote.http11.http.request.HttpRequest;
+import org.apache.coyote.http11.http.response.HttpResponse;
 
 public class HomeHandler implements Handler {
     private static final String DEFAULT_MESSAGE = "Hello world!";
