+++ tomcat/src/main/java/org/apache/coyote/http11/http/request/HttpMethod.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.request;
+package org.apache.coyote.http11.http.request;
 
 import java.util.Arrays;
 import java.util.NoSuchElementException;
