+++ tomcat/src/main/java/org/apache/coyote/http11/http/response/HttpStatus.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.response;
+package org.apache.coyote.http11.http.response;
 
 public enum HttpStatus {
     OK(200, "OK");
